requirements.txt
fastapi
uvicorn
jinja2
pydantic
pyyaml
tabulate
